[FIX] Change comparing Long and int #86

Closes #86

See merge request s07-final/S07P31A504!90

+++ server/ore/src/main/java/io/ssafy/p/k7a504/ore/teamUser/repository/TeamUserRepository.java
@@ -15,20 +15,20 @@ public interface TeamUserRepository extends JpaRepository<TeamUser, Long> {
     Slice<TeamUser> findByTeamId(Long teamId, Long userId, Pageable pageable);
     Optional<TeamUser> findByUserIdAndTeamId(Long userId, Long teamId);
     @Query(value = "SELECT count(*) FROM team_user WHERE user_id In :list AND team_id = :teamId", nativeQuery = true)
-    Long countByUserIdListAndTeamId(@Param("list")List<Long> userIdList, @Param("teamId")Long teamId);
+    int countByUserIdListAndTeamId(@Param("list")List<Long> userIdList, @Param("teamId")Long teamId);
     @Query("SELECT tu FROM TeamUser tu JOIN FETCH User u On tu.user.id = u.id WHERE u.name LIKE %?1%  AND tu.team.id = ?2 AND u.id != ?3 AND tu.user.role != 'OWNER'")
     Slice<TeamUser> findTeamUserByUserNameAndTeamId(String name, Long teamId, Long userId, Pageable pageable);
     @Query("SELECT tu FROM TeamUser tu JOIN FETCH User u On tu.user.id = u.id WHERE u.name LIKE %?1% AND tu.team.id = ?2 AND u.id != ?3 AND tu.user.role != 'OWNER'")
     Slice<TeamUser> findTeamUserByUserNicknameAndTeamId(String nickname, Long teamId, Long userId, Pageable pageable);
     @Query(value = "SELECT count(*) FROM team_user WHERE id In :list AND team_id = :teamId", nativeQuery = true)
-    Long countByTeamUserIdListAndTeamId(@Param("list")List<Long> teamUserIdList, @Param("teamId")Long teamId);
+    int countByTeamUserIdListAndTeamId(@Param("list")List<Long> teamUserIdList, @Param("teamId")Long teamId);
 
     @Query(value = "SELECT count(*) FROM team_user WHERE id In :list AND team_user.role IN ('OWNER','LEADER')", nativeQuery = true)
-    Long countRoleOverLeader(@Param("list")List<Long> teamUserIdList);
+    int countRoleOverLeader(@Param("list")List<Long> teamUserIdList);
 
     @Query(value = "SELECT count(*) FROM team_user WHERE id In :list  AND team_user.role IN ('OWNER','LEADER', 'MANAGER')", nativeQuery = true)
-    Long countRoleOverManager(@Param("list")List<Long> teamUserIdList);
+    int countRoleOverManager(@Param("list")List<Long> teamUserIdList);
 
     @Query(value = "SELECT count(*) FROM team_user WHERE id In :list", nativeQuery = true)
-    Long countById(@Param("list")List<Long> userIdList);
+    int countById(@Param("list")List<Long> userIdList);
 }
